Fixing the missing arrow for font family combo in firefox (again)
As reported in, jgraph/mxgraph#343 the drop down menus are not
visible for font name. Two statements conditioned on firefox
a made -18px top margin adjustments causing the arrow to depart
from the visible bounds.

The results of git blame yielded mild amusement. This change
reverts two prior bugfixes from 2017 and 2016. One change was
described as "[fixing the] missing arrow for font family
combo in firefox".

  jgraph/mxgraph2@af3bc081ab843470804dcf4933c311e424dd119c
  jgraph/mxgraph2@88d08fbd28eba51bb82e741171f032cfa8549240