[WIP] Remove geBtn class on add property button
The INPUT element is not stylized so this leads to an inconsistency
between the button and the input element. Removing the class means the
input field is more or less aligned with the button.

Although having input and button styles aligned is nice, a more
pragmatic explaination can be found when inspecting the layout.
The .geBtn makes a bigger button and overflows the parent elemnent.

Either...
1) The parent should expand to accomodate the size of the button
  or
2) Remove the style and the two unstyleized elements be  similarly
   sized.

I took the latter lazy option. It's also looks better.

Related to jgraph/mxgraph#343